\name{orthoplanes}
\alias{orthoplanes}
\title{
Calculate Vec_Penp, landmark 0 and the vectors between landmarks 0-2, 1-1_Line, 5-6 and 7-8.
}
\description{
This function calculates the normalized coordinates of a vector perpendicular to the longitudinal (anteroposterior, in anatomical terms) axis of the plane defined by landmarks 1, 2, and 1_Line. This vector is referred to as Vec_Penp.
This function also computes the vectors formed by landmarks 2 and 0 (Vec_0_2), landmarks 1 and 1_Line (Vec_1_1Line), landmarks 5 (renamed as CS1B) and 6 (Vec_CS1) and landmarks 7 (renamed as CS2B) and 8 (Vec_CS2).
This function also calculates the X, Y and Z coordinates of a new landmark (0), defined as the projection of landmark 2 onto the vector formed by landmarks 1 and 1_Line.
This function also transform the result_matrix used as input by modifying the order of its rows and renaming the landmarks.
}
\usage{
orthoplanes(result_matrix)
}
\arguments{
  \item{result_matrix}{
must be a 12x3 matrix where the landmarks 1, 2, 3, 4, 1_Line, 5, 6, 7, 8, 9, 10 and 11 are arranged from top to bottom in that same order. This matrix can be generated by calling the function mirror_LM1, so the execution of the latter is required to run this function.
}
}
\details{
This function can only be executed with matrices containing three columns; i.e., with three-dimensional points where the coordinates for X, Y, and Z are each located in a separate column).
The reason for including two functions that are so similar (this one and pMV) in the package is that when designing the protocol, a function should be included to allow following the described workflow in the standard protocol (protocols.io). However, in our personal research case, the workflow was slightly different, so we had to desing a new function that addressed these circumstances (pMV).
}
\value{
A new result_matrix with 18 rows, containing in the following order the three-dimensional coordinates of: LM1, LM2, LM3, LM4, LM1_Line, CS1B (formerly LM5), Vec_CS1, CS2B (formerly LM7), Vec_CS2, Vec_Penp, Vec_0_2, Vec_1_1Line, LM0, LM6, LM8, LM9, LM10, and LM11.
Besides, the function will generate four new objects in the R enviroment: "Vec_0_2", "Vec_CS1", "Vec_CS2" and "Vec_Penp"
}
\references{
Protocols.io - to be published
}
\author{
Alejandro Martín-Sánchez
}
\note{
As a low-level user of R, while I grasp the logic of the language and handle its fundamental principles, I have turned to AI to generate the code for the entire package in order to save hours of work. A seasoned developer or even someone proficient in the field may have easily noticed this, as the code may not be as efficient as it could be. For this reason, any suggestions on how to improve it are more than welcome :)
}
\examples{
require(MandOri)
text <- "-3.176e+01 -7.557e+00 -1.441e+02 -1.478e+00 6.264e+00 -1.046e+02 -5.153e-02 -2.313e+01 -9.379e+01 -5.700e-01 -1.662e+01 -1.083e+02 -2.924e+01 -5.551e+00 -1.335e+02 -2.191e+01 -5.099e+00 -1.348e+02 -2.557e+01 -1.035e+00 -1.221e+02 -1.980e+01 -6.242e-01 -1.243e+02 -4.616e+01 -3.496e+01 -1.652e+02 -1.790e+01 2.789e+00 -1.106e+02 -4.694e+01 1.235e+01 -2.013e+02"
result_matrix <- getmatrix(text)
mirror_LM1(result_matrix)
orthoplanes(result_matrix)
}
