\name{point_M}
\alias{point_M}
\title{
Calculate M and the vector MB.
}
\description{
This function calculates the X, Y and Z coordinates of a new point (M), defined as the projection of point B onto the vector formed by points A and A_Prime. This function also computes the vector formed by points B and M.
}
\usage{
point_M(x)
}
\arguments{
  \item{x}{
x must be a 5x3 matrix where the points A, B, C, D and A_Prime are arranged from top to bottom in that same order. This matrix (result_matrix_A_Prime) could be generated by calling the function mirror_A, so the execution of the latter is required to run this function.
}
}
\details{
This function can only be executed with matrices containing three columns; i.e., with three-dimensional points where the coordinates for X, Y, and Z are each located in a separate column)
}
\value{
Three-dimentional coordinates of the new calculated point
The values of the vector MB
Besides, the function will generate a new object (point_M_coordinates) that includes the coordinates of the new point M.
}
\references{
Protocols.io - to be published
}
\author{
Alejandro Martín-Sánchez
}
\note{
As a low-level user of R, while I grasp the logic of the language and handle its fundamental principles, I have turned to AI to generate the code for the entire package in order to save hours of work. A seasoned developer or even someone proficient in the field may have easily noticed this, as the code may not be as efficient as it could be. For this reason, any suggestions on how to improve it are more than welcome :)
}
\examples{
require(MandiOri)
text <- "2 2 8 5 8 4 6 9 7 6 3 2"
result_matrix <- text2matrix(text)
mirror_A(result_matrix)
point_M(result_matrix_A_Prime)
}
