\name{point_M_Vs}
\alias{point_M_Vs}
\title{
Calculate point M and the vectors MB and AA_Line
}
\description{
This function calculates the X, Y and Z coordinates of a new point (M), defined as the projection of point B onto the vector formed by points A and A_Line. This function also computes the vectors formed by points B and M and points A and A_Line.
}
\usage{
point_M_Vs(x)
}
\arguments{
  \item{x}{
must be a 10x3 matrix where the points A, B, C, D, "Gonion", "CP3", "Cond_Midpoint", "Ramus_Post", "Ramus_Root" and A_Line are arranged from top to bottom in that same order. This matrix can be generated by calling the function mirror_A, so the execution of the latter is required to run this function.
}
}
\details{
This function can only be executed with matrices containing three columns; i.e., with three-dimensional points where the coordinates for X, Y, and Z are each located in a separate column).
The reason for including two functions that are so similar (this one and pMV) in the package is that when designing the protocol, a function should be included to allow following the described workflow in the standard protocol (protocols.io). However, in our personal research case, the workflow was slightly different, so we had to desing a new function that addressed these circumstances (pMV).
}
\value{
Three-dimentional coordinates of the new calculated point M.
The values of the vectors MB and AA_Line.
Besides, the function will generate two new objects in the R enviroment: "point_M_coordinates" (1) that includes the coordinates of the new point M and "result_matrix" (2), the matrix previously entered as input but with three new rows added in this order: coordinates of point M, MB vector, and AA_Line vector.
}
\references{
Protocols.io - to be published
}
\author{
Alejandro Martín-Sánchez
}
\note{
As a low-level user of R, while I grasp the logic of the language and handle its fundamental principles, I have turned to AI to generate the code for the entire package in order to save hours of work. A seasoned developer or even someone proficient in the field may have easily noticed this, as the code may not be as efficient as it could be. For this reason, any suggestions on how to improve it are more than welcome :)
}
\examples{
require(MandOri)
text <- "2 2 8 5 8 4 6 9 7 6 3 2 5 9 8 5 4 8 5 9 4 5 4 7 4 5 8"
result_matrix <- getmatrix(text)
mirror_A(result_matrix)
point_M_Vs(result_matrix)
}
