\name{CalcEstDistances}
\alias{CalcEstDistances}
\title{
Calculate linear distances in mm from a given set of three-dimentional points
}
\description{
Calculate linear distances (in millimetres) that will subsequently be used for size standardization processes.The difference with distR is that this function utilizes the set of points described in the standard protocol (protocols.io), which has a specific order and size (coordiantes of 13 points entered as matrix).
}
\usage{
CalcEstDistances(x)
}
\arguments{
  \item{x}{
must be a 13x3 matrix containing all the coordinates of the points to be included in the calculations. This matrix can be generated by running the functions getmatrix (first), mirror_A (second), and point_M_Vs (third). For further details, please refer to the link provided in the "References" section.
}
}
\details{
This function can only be executed with matrices containing three columns; i.e., with three-dimensional points where the coordinates for X, Y, and Z are each located in a separate column).
The reason for including two functions that are so similar (this one and distR) in the package is that when designing the protocol, a function should be included to allow following the described workflow in the standard protocol (protocols.io). However, in our personal research case, the workflow was slightly different because we already had the coordinates of the points stored in a different way, so we had to desing a new function that addressed these circumstances (distR).
}
\value{
Distance in milimeters of:
- Mandibular corpus length
- Mandibular ramuss lateral length
- Mandibular ramus width
- Dental arch breadth
For further details on the measurements and the reason for their calculation, see references (see below).
}
\references{
Protocols.io - to be published
}
\author{
Alejandro Martín-Sánchez
}
\note{
As a low-level user of R, while I grasp the logic of the language and handle its fundamental principles, I have turned to AI to generate the code for the entire package in order to save hours of work. A seasoned developer or even someone proficient in the field may have easily noticed this, as the code may not be as efficient as it could be. For this reason, any suggestions on how to improve it are more than welcome :)
}
\examples{
require(MandOri)
text <- "2 2 8 5 8 4 6 9 7 6 3 2 5 9 8 5 4 8 5 9 4 5 4 7 4 5 8"
result_matrix <- getmatrix(text)
mirror_A(result_matrix)
point_M_Vs(result_matrix)
CalcEstDistances(result_matrix)
}
